fix: Location을 Position으로 renamed
