+++ src/main/java/mission/controller/Place.java
@@ -1,10 +1,7 @@
 package mission.controller;
 
-import mission.constants.Text;
-
 public record Place(
     int id,
     String name,
     String address
-) {
-}
+) { }
